lab cleanup
