+++ htk/lab.go
@@ -10,19 +10,82 @@ import (
 	"strings"
 )
 
-// An annotation has a starting and ending time in seconds, with a text label.
-type Annotation struct {
-	start float64
-	end   float64
-	label string
-}
-
 // A lab contains a collection of annotations.
 // The HTK Label format is defined at http://www.seas.ucla.edu/spapl/weichu/htkbook/node113_mn.html
 type Lab struct {
 	annotations []Annotation
 	name        string
 	precision   uint8
+}
+
+// Sets the annotations of a lab object
+func (lab *Lab) SetAnnotations(annotations []Annotation) {
+	lab.annotations = annotations
+}
+
+// Gets the annotations of a lab object
+func (lab Lab) GetAnnotations() []Annotation {
+	return lab.annotations
+}
+
+// Pushes an annotation to the end of a lab object
+func (lab *Lab) PushAnnotation(annotation Annotation) {
+	lab.annotations = append(lab.annotations, annotation)
+}
+
+// Appends an annotations object to the end of a lab object
+func (lab *Lab) AppendAnnotations(annotations []Annotation) {
+	lab.annotations = append(lab.annotations, annotations...)
+}
+
+// Removes all annotations from a lab object
+func (lab *Lab) ClearAnnotations() {
+	lab.annotations = []Annotation{}
+}
+
+// Dumps all labels in a lab object to a slice
+func (lab *Lab) DumpLabels() []string {
+	var result []string
+
+	for _, annotation := range lab.annotations {
+		result = append(result, annotation.label)
+	}
+
+	return result
+}
+
+// Gets the name of a lab object
+func (lab Lab) GetName() string {
+	return lab.name
+}
+
+// Sets the name of a lab object
+func (lab *Lab) SetName(name string) {
+	lab.name = name
+}
+
+// Gets the precision of a lab object
+func (lab Lab) GetPrecision() uint8 {
+	return lab.precision
+}
+
+// Sets the precision of a lab object
+func (lab *Lab) SetPrecision(precision uint8) {
+	lab.precision = precision
+}
+
+// Gets the total duration of a lab by getting the difference in global start and end.
+func (lab Lab) GetDuration() (result float64) {
+	// calculate using start and end in case lab file doesnt start at 0
+	start := lab.annotations[0].start
+	end := lab.annotations[len(lab.annotations)-1].end
+
+	return end - start
+}
+
+// Gets the total amount of annotations in a lab file.
+func (lab Lab) GetLength() (result int) {
+	return len(lab.annotations)
 }
 
 // Takes a path to a .lab file and reads its contents into a Lab.
@@ -108,7 +171,7 @@ func (lab Lab) WriteLab(path string, overwrite ...bool) {
 	} else {
 		// if the path is a file, check if it already exists and if overwrite is false
 		if pathInfo != nil && !overwrite[0] {
-			log.Fatalf("file %s already exists", path)
+			log.Fatalf("error: file %s already exists", path)
 		}
 		fileName = path
 	}
